+++ lib/Conversion/TritonInstrumentToLLVM/CMakeLists.txt
@@ -8,5 +8,5 @@ add_triton_library(TritonInstrumentToLLVM
     TritonGPUIR
     TritonInstrumentIR
     TritonNvidiaGPUIR
-    NVGPUIR
+    NVGIR
 )
