+++ test/Conversion/tritoninstrument_to_llvm.mlir
@@ -4,7 +4,7 @@
 
 module attributes {"ttg.num-warps" = 4 : i32, ttg.target = "cuda:90"} {
 // CHECK-LABEL: @experimental_buffer_pointers_tmem
-// CHECK:nvgpu.tensor_memory_base
+// CHECK:nvg.tensor_memory_base
 tt.func private @experimental_buffer_pointers_tmem() {
   tti.experimental_buffer_pointers [0, 42], tensor_mem : tensor<2xi64, #blocked>
   tt.return
