+++ test/Hopper/WarpSpecialization/ws_task_id_propagation.mlir
@@ -1,4 +1,4 @@
-// RUN: triton-opt %s -split-input-file --nvgpu-test-taskid-propagate=num-warp-groups=2 | FileCheck %s
+// RUN: triton-opt %s -split-input-file --nvg-test-taskid-propagate=num-warp-groups=2 | FileCheck %s
 
 #blocked = #ttg.blocked<{sizePerThread = [1, 1], threadsPerWarp = [1, 32], warpsPerCTA = [2, 2], order = [1, 0]}>
 #blocked1 = #ttg.blocked<{sizePerThread = [1, 1], threadsPerWarp = [1, 32], warpsPerCTA = [1, 4], order = [1, 0]}>
